Refactor BallDontLie source: schema inference, replace strategy, rename scheme to balldontlie
- Use schema inference (KnownSchema=false) instead of predefined column
  lists; drop all flatten* helpers so nested objects pass through as raw
  JSON columns
- All tables use the replace strategy: the API exposes no time-interval
  filters, so every run is a full season fetch; HandlesIncrementality
  returns true (the source owns incrementality via full replace)
- Stream one Arrow batch per page instead of buffering all pages
- Add a rate limiter for the free-tier 5 req/min limit
- Rename URI scheme balldontlie-fifa -> balldontlie, and rename the
  package/struct/dir/docs accordingly for consistency
